requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
python-dotenv
aiofiles